Added AdminCotroller Added CarDAO and MopedDAO with relative insert functions. Old UserController is now called LoginController.

+++ src/DomainModel/Car.java
@@ -4,7 +4,7 @@ public class Car extends Vehicle {
     private int nseats;
     private RoadsideAssistance assistance;
 
-    public Car(String model, String plate, int daily_price, boolean available, int nseats,
+    public Car(String model, String plate, float daily_price, boolean available, int nseats,
                RoadsideAssistance assistance) {
         super(model, plate, daily_price, available);
         this.nseats = nseats;
