+++ src/DomainModel/Vehicle.java
@@ -2,10 +2,10 @@ package DomainModel;
 
 public abstract class Vehicle {
     protected String model, plate;
-    protected int daily_price;
+    protected float daily_price;
     protected boolean available;
 
-    public Vehicle(String model, String plate, int daily_price, boolean available) {
+    public Vehicle(String model, String plate, float daily_price, boolean available) {
         this.model = model;
         this.plate = plate;
         this.daily_price = daily_price;
@@ -20,7 +20,7 @@ public abstract class Vehicle {
         return plate;
     }
 
-    public int getDaily_price() {
+    public float getDaily_price() {
         return daily_price;
     }
